requirements.txt

Flask==3.0.0
playwright==1.40.0
Flask-JWT-Extended==4.6.0
PyJWT==2.8.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
authlib==1.3.0
itsdangerous==2.1.2
werkzeug==3.0.1
python-dotenv==1.0.0
requests==2.31.0
Jinja2==3.1.2
MarkupSafe==2.1.3
click==8.1.7
blinker==1.7.0
